Remove WhileMessageNode which is not used any longer
But make sure there's a primitive to fallback if needed.
