+++ src/som/primitives/ast/block_primitives.py
@@ -1,15 +1,4 @@
-from som.primitives.primitives import Primitives
-from som.vmobjects.primitive import Primitive
+from som.primitives.block_primitives import BlockPrimitivesBase as _Base
 
 
-def _restart(ivkbl, rcvr, args):
-    raise RuntimeError(
-        "Restart primitive is not supported, #whileTrue: "
-        "and #whileTrue: are intrisified so that #restart "
-        "is not needed."
-    )
-
-
-class BlockPrimitives(Primitives):
-    def install_primitives(self):
-        self._install_instance_primitive(Primitive("restart", _restart))
+BlockPrimitives = _Base
